ULTRA MEGA MODÜLER
Bina inşası için çözüm seçeneği yazıldı.
