+++ Sorunlar.txt
@@ -1,3 +1,1 @@
 Movement -> player_movement() -> Neden move_to kullanıldığını anlamadım.
-Çapraz hareketlerde bazen hücre geç çöküyor.
-destroy_tiles, spatial_index ile ilişkilendirilebilir ?
