+++ çözümler.txt
@@ -1,0 +1,1 @@
+BİNA İNŞASI ÇOK UZAKLARDA GERÇEKLEŞTİRİLEBİLİR.
